1. Minor restructuring.
