+++ iris-mpc-utils/src/graphs/plaintext/mod.rs
@@ -1,4 +1,4 @@
 pub mod builder;
 pub mod py_bindings;
 pub mod store;
-pub mod test_utils;
+pub mod utils;
