+++ iris-mpc-utils/src/state/mod.rs
@@ -1,4 +1,3 @@
 pub mod aws;
 pub mod fsys;
-pub mod imem;
 pub mod pgres;
